requirements.txt
asgiref==3.7.2
Django==3.1.5
django-dotenv==1.4.2
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.6.3
